yutuval/requirements.txt
py-cord
python-dotenv
aiosqlite
flask
